[omp] [call spots] duplicate spot detection is not repeated, closes #124
