+++ coppafish/call_spots/base.py
@@ -1,58 +1,5 @@
-from scipy.spatial import KDTree
-from scipy.sparse.linalg import svds
 import numpy as np
-from typing import List
-
-from .. import log
-
-
-def get_non_duplicate(
-    tile_origin: np.ndarray, use_tiles: List, tile_centre: np.ndarray, spot_local_yxz: np.ndarray, spot_tile: np.ndarray
-) -> np.ndarray:
-    """
-    Find duplicate spots as those detected on a tile which is not tile centre they are closest to.
-
-    Args:
-        tile_origin: `float [n_tiles x 3]`.
-            `tile_origin[t,:]` is the bottom left yxz coordinate of tile `t`.
-            yx coordinates in `yx_pixels` and z coordinate in `z_pixels`.
-            This is saved in the `stitch` notebook page i.e. `nb.stitch.tile_origin`.
-        use_tiles: ```int [n_use_tiles]```.
-            Tiles used in the experiment.
-        tile_centre: ```float [3]```
-            ```tile_centre[:2]``` are yx coordinates in ```yx_pixels``` of the centre of the tile that spots in
-            ```yxz``` were found on.
-            ```tile_centre[2]``` is the z coordinate in ```z_pixels``` of the centre of the tile.
-            E.g. for tile of ```yxz``` dimensions ```[2048, 2048, 51]```, ```tile_centre = [1023.5, 1023.5, 25]```
-            Each entry in ```tile_centre``` must be an integer multiple of ```0.5```.
-        spot_local_yxz: ```int [n_spots x 3]```.
-            Coordinates of a spot s on tile spot_tile[s].
-            ```yxz[s, :2]``` are the yx coordinates in ```yx_pixels``` for spot ```s```.
-            ```yxz[s, 2]``` is the z coordinate in ```z_pixels``` for spot ```s```.
-        spot_tile: ```int [n_spots]```.
-            Tile each spot was found on.
-
-    Returns:
-        ```bool [n_spots]```.
-            Whether spot_tile[s] is the tile that spot_global_yxz[s] is closest to.
-    """
-    assert spot_local_yxz.shape[0] == spot_tile.shape[0]
-    tile_centres = tile_origin[use_tiles] + tile_centre
-    # Do not_duplicate search in 2D as overlap is only 2D
-    tree_tiles = KDTree(tile_centres[:, :2])
-    if np.isnan(tile_origin[np.unique(spot_tile)]).any():
-        nan_tiles = np.unique(spot_tile)[np.unique(np.where(np.isnan(tile_origin[np.unique(spot_tile)]))[0])]
-        log.error(
-            ValueError(
-                f"tile_origin for tiles\n{nan_tiles}\ncontains nan values but some spot_tile "
-                f"also contains these tiles. Maybe remove these from use_tiles to continue.\n"
-                f"Also, consider coppafish.plot.n_spots_grid to check if these tiles have few spots."
-            )
-        )
-    spot_global_yxz = spot_local_yxz + tile_origin[spot_tile]
-    all_nearest_tile_ind = tree_tiles.query(spot_global_yxz[:, :2])[1]
-    not_duplicate = np.asarray(use_tiles)[all_nearest_tile_ind.flatten()] == spot_tile
-    return not_duplicate
+import scipy
 
 
 def bayes_mean(
@@ -120,7 +67,7 @@ def compute_bleed_matrix(
             dye_d_colours.append(spot_colours[relevant_gene_mask, r, :])
         # now we have all the good colours for dye d, compute the SVD
         dye_d_colours = np.concatenate(dye_d_colours, axis=0)
-        u, s, v = svds(dye_d_colours, k=1)
+        u, s, v = scipy.sparse.linalg.svds(dye_d_colours, k=1)
         v = v[0]
         # make sure largest entry in v is positive
         v *= np.sign(v[np.argmax(np.abs(v))])
