+++ coppafish/call_spots/test/test_call_spots_base.py
@@ -3,21 +3,6 @@ import numpy as np
 from coppafish.call_spots import base
 
 
-def test_get_non_duplicate():
-    # Place two tiles
-    tile_origin = np.array([[0,0,0],[100,0,0]])
-    # Use both tiles
-    use_tiles = [0,1]
-    tile_centre = np.array([50,0,0], float)
-    spot_local_yxz = np.array([[80,0,0], [110,0,0]], int)
-    # Say each spot was found on the first tile
-    spot_tile = np.array([0,0], int)
-    output = base.get_non_duplicate(tile_origin, use_tiles, tile_centre, spot_local_yxz, spot_tile)
-    assert spot_local_yxz.shape[0] == output.size, 'Expect output to have the same number of spots'
-    assert output[0], 'First spot should be found closest to the first tile'
-    assert not output[1], 'Second spot should be found closest to the other tile'
-
-
 def test_bayes_mean():
     seed = 0
     np.random.seed(seed)
@@ -31,11 +16,11 @@ def test_bayes_mean():
     spot_colours[:, 0] *= 10
     bayes_mean = base.bayes_mean(spot_colours, expected_colour, 0.1, 50)
     # check shape
-    assert bayes_mean.shape == (n_channels,), 'Expect output to have the same number of channels'
+    assert bayes_mean.shape == (n_channels,), "Expect output to have the same number of channels"
     # check that the first entry has average approximately 10
-    assert np.isclose(bayes_mean[0], 10, atol=1), 'Expect first column to have average 10'
+    assert np.isclose(bayes_mean[0], 10, atol=1), "Expect first column to have average 10"
     # check that the other entries are approximately 0
-    assert np.all(np.isclose(bayes_mean[1:], 0, atol=0.1)), 'Expect other columns to have average 0'
+    assert np.all(np.isclose(bayes_mean[1:], 0, atol=0.1)), "Expect other columns to have average 0"
 
 
 def test_compute_bleed_matrix():
@@ -51,14 +36,15 @@ def test_compute_bleed_matrix():
     # gene assignments
     gene_no = np.random.randint(0, n_genes, n_spots)
     # spot colours
-    spot_colours = (expected_bled_codes[gene_no]
-                    + np.random.normal(0, 0.1, (n_spots, n_rounds, n_channels)))
-    bleed_matrix = base.compute_bleed_matrix(spot_colours=spot_colours, gene_no=gene_no,
-                                             gene_codes=gene_codes, n_dyes=n_dyes)
+    spot_colours = expected_bled_codes[gene_no] + np.random.normal(0, 0.1, (n_spots, n_rounds, n_channels))
+    bleed_matrix = base.compute_bleed_matrix(
+        spot_colours=spot_colours, gene_no=gene_no, gene_codes=gene_codes, n_dyes=n_dyes
+    )
     # check shape
-    assert bleed_matrix.shape == (n_dyes, n_channels), 'Expect output to have the same shape as the bleed matrix'
+    assert bleed_matrix.shape == (n_dyes, n_channels), "Expect output to have the same shape as the bleed matrix"
     # check that the bleed matrix is close to the expected bleed matrix
     # need to normalise each row of expected_bleed first
     expected_bleed = expected_bleed / np.sum(expected_bleed, axis=1)[:, None]
-    assert np.all(np.isclose(bleed_matrix, expected_bleed, atol=0.1)), \
-        'Expect bleed matrix to be close to the expected bleed matrix'
+    assert np.all(
+        np.isclose(bleed_matrix, expected_bleed, atol=0.1)
+    ), "Expect bleed matrix to be close to the expected bleed matrix"
