+++ coppafish/omp/scores.py
